pyproject.toml
[tool.poetry]
name = "tylaireum"
version = "0.1.0"
description = ""
authors = ["Bryan Hu <bryan.hu.2020@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Jinja2 = "^2.11.3"
lark-parser = "^0.11.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.812"
pylint = "^2.7.2"
black = {version = "^20.8b1", allow-prereleases = true}

[tool.poe.tasks]
test = 'pytest -vvv'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
